:white_check_mark: migration des fonction de screen a simulation
